Legger til datovelger og nødvendig funksjonalitet+
- Deler opp og flytter omDeg typer som IMedlemskap og ISivilstatus inn til egne ts-filer
