+++ src/mock/person.json
@@ -12,7 +12,7 @@
     "innvandretDato": "1991-12-28",
     "utvandretDato": "",
     "oppholdstillatelse": "",
-    "sivilstand": "UGIF",
+    "sivilstand": "GIFT",
     "språk": "NB",
     "statsborgerskap": "NOR",
     "privattelefon": "",
