--- src/models/steg/omDeg.ts
@@ -1,26 +1,0 @@
-import { IBooleanFelt, IDatoFelt, ITekstFelt } from '../søknadsfelter';
-import { IPeriode } from '../søknad';
-
-export interface ISivilstatus {
-  søkerHarSøktSeparasjon?: IBooleanFelt;
-  datoSøktSeparasjon?: IDatoFelt;
-  søkerGiftIUtlandet?: IBooleanFelt;
-  søkerSeparertEllerSkiltIUtlandet?: IBooleanFelt;
-  begrunnelseForSøknad?: ITekstFelt;
-  datoForSamlivsbrudd?: IDatoFelt;
-  datoFlyttetFraHverandre?: IDatoFelt;
-  datoEndretSamvær?: IDatoFelt;
-  begrunnelseAnnet?: ITekstFelt;
-}
-
-export interface IMedlemskap {
-  søkerOppholderSegINorge?: IBooleanFelt;
-  søkerBosattINorgeSisteTreÅr?: IBooleanFelt;
-  perioderBoddIUtlandet?: IUtenlandsopphold[];
-}
-
-export interface IUtenlandsopphold {
-  react_key: string;
-  periode: IPeriode;
-  begrunnelse: ITekstFelt;
-}
