+++ src/models/søknad.ts
@@ -1,10 +1,11 @@
 import { IAktivitet } from './steg/aktivitet/aktivitet';
 import { IPerson } from './person';
 import { IVedlegg } from './vedlegg';
-import { IMedlemskap, ISivilstatus } from './steg/omDeg';
 import { IBooleanFelt, IDatoFelt } from './søknadsfelter';
 import { IBosituasjon } from './steg/bosituasjon';
 import { IDinSituasjon } from './steg/dinsituasjon/meromsituasjon';
+import { ISivilstatus } from './steg/omDeg/sivilstatus';
+import { IMedlemskap } from './steg/omDeg/medlemskap';
 
 export interface ISøknad {
   bekreftet?: boolean;
