+++ src/søknad/steg/1-omdeg/medlemskap/PeriodeBoddIUtlandet.tsx
@@ -9,9 +9,9 @@ import FeltGruppe from '../../../../components/gruppe/FeltGruppe';
 import Utenlandsopphold from './Utenlandsopphold';
 import { dagensDato } from '../../../../utils/dato';
 import subDays from 'date-fns/subDays';
-import { IUtenlandsopphold } from '../../../../models/steg/omDeg';
 import { hentTekst } from '../../../../utils/søknad';
 import { hentUid } from '../../../../utils/uuid';
+import { IUtenlandsopphold } from '../../../../models/steg/omDeg/medlemskap';
 
 const PeriodeBoddIUtlandet: FC = () => {
   const { søknad, settSøknad } = useSøknadContext();
