+++ src/søknad/steg/1-omdeg/medlemskap/Utenlandsopphold.tsx
@@ -6,9 +6,9 @@ import { Undertittel } from 'nav-frontend-typografi';
 import classnames from 'classnames';
 import SlettKnapp from '../../../../components/knapper/SlettKnapp';
 import { hentTittelMedNr } from '../../../../language/utils';
-import { IUtenlandsopphold } from '../../../../models/steg/omDeg';
 import PeriodeDatovelgere from '../../../../components/dato/PeriodeDatovelger';
 import { hentTekst } from '../../../../utils/søknad';
+import { IUtenlandsopphold } from '../../../../models/steg/omDeg/medlemskap';
 
 interface Props {
   utenlandsopphold: IUtenlandsopphold;
