+++ src/søknad/steg/1-omdeg/sivilstatus/SøkerHarSøktSeparasjon.tsx
@@ -7,24 +7,21 @@ import useSøknadContext from '../../../../context/SøknadContext';
 import FeltGruppe from '../../../../components/gruppe/FeltGruppe';
 import LocaleTekst from '../../../../language/LocaleTekst';
 import KomponentGruppe from '../../../../components/gruppe/KomponentGruppe';
+import NårFlyttetDereFraHverandre from './begrunnelse/NårFlyttetDereFraHverandre';
 
 interface Props {
   settDato: (date: Date | null, objektnøkkel: string, tekst: string) => void;
 }
 const SøkerHarSøktSeparasjon: React.FC<Props> = ({ settDato }) => {
   const { søknad } = useSøknadContext();
-  const { sivilstatus } = søknad;
+  const { datoSøktSeparasjon, datoFlyttetFraHverandre } = søknad.sivilstatus;
   return (
     <KomponentGruppe>
       <Datovelger
         settDato={(e) =>
           settDato(e, 'datoSøktSeparasjon', 'sivilstatus.separasjon.datosøkt')
         }
-        valgtDato={
-          sivilstatus.datoSøktSeparasjon
-            ? sivilstatus.datoSøktSeparasjon.verdi
-            : undefined
-        }
+        valgtDato={datoSøktSeparasjon ? datoSøktSeparasjon.verdi : undefined}
         tekstid={'sivilstatus.separasjon.datosøkt'}
         datobegrensning={DatoBegrensning.TidligereDatoer}
       />
@@ -33,6 +30,12 @@ const SøkerHarSøktSeparasjon: React.FC<Props> = ({ settDato }) => {
           <LocaleTekst tekst={'sivilstatus.somgift'} />
         </AlertStripeInfo>
       </FeltGruppe>
+      {datoSøktSeparasjon && (
+        <NårFlyttetDereFraHverandre
+          settDato={settDato}
+          datoFlyttetFraHverandre={datoFlyttetFraHverandre}
+        />
+      )}
     </KomponentGruppe>
   );
 };
